requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
pillow
pyyaml
tqdm
matplotlib
pydicom
